server: modernize websockets
Move off Offer/Broker, preferring AsyncStream.

In this model, servers receive an AsyncStream representing the frames of
an incoming WebSocket client. The server responds with an AsyncStream of
frames to be written to the client. When the response ends, the
dispatcher knows to close the transport. This has the effect of
discarding incoming frames.